Optimised speed & memory by ~8x by moving from uint8_t vectors to uint64_t vectors to compact edge selections i.e, 64 edge selections can now be represented in a single 8-byte value (i.e, 1 bit per edge) rather than before where it was 1 byte per edge (7 wasted bits).

+++ include/solvers/brute_force_solver.h
@@ -18,7 +18,7 @@ public:
 private:
     double backtrack(const Graph& graph, 
                      const ExpectedRequests& requests, 
-                     std::vector<uint8_t>& selected_edges, 
+                     std::vector<uint64_t>& selected_edges, 
                      std::size_t index);
 };
